+++ engine/core/tests/BF_SHA_HelperCoverage.cpp
@@ -63,7 +63,7 @@ int TestSmallSigma0_RewriteEmitEvalConsistency() {
 
     const auto eval = Eval::EvaluateConstant(rewritten, 32);
     BF_TEST(eval.status == Eval::EvalStatus::Success);
-    BF_TEST(static_cast<uint32_t>(eval.value) == (RotR32(x, 7) ^ RotR32(x, 18) ^ (x >> 3)));
+    BF_TEST(eval.value == (RotR32(x, 7) ^ RotR32(x, 18) ^ (x >> 3)));
 
     auto sigmaVar = b.SmallSigma0(b.Var());
     auto emitted = Codegen::EmitCExpr(MakeShaSafeEngine().Rewrite(sigmaVar), 32);
@@ -81,7 +81,7 @@ int TestSmallSigma1_RewriteEmitEvalConsistency() {
 
     const auto eval = Eval::EvaluateConstant(rewritten, 32);
     BF_TEST(eval.status == Eval::EvalStatus::Success);
-    BF_TEST(static_cast<uint32_t>(eval.value) == (RotR32(x, 17) ^ RotR32(x, 19) ^ (x >> 10)));
+    BF_TEST(eval.value == (RotR32(x, 17) ^ RotR32(x, 19) ^ (x >> 10)));
 
     auto sigmaVar = b.SmallSigma1(b.Var());
     auto emitted = Codegen::EmitCExpr(MakeShaSafeEngine().Rewrite(sigmaVar), 32);
